requirements.txt
streamlit==1.39.0
pandas==2.2.2
openpyxl==3.1.5
xlsxwriter==3.2.0
pdfplumber==0.11.4
reportlab==4.2.5
pyinstaller==6.6.0
google-generativeai
python-dotenv
supabase
python-dateutil
